Proof of Concept to Continuously evaluate visited artifacts.
This is just a quick and (very dirty) implementation to present the idea that we can continuously evaluate a library seen on users browser, without the need of the click on the extension icon on top right corner.
We can also update the icon with the highest security score.

This implementation is for Maven only and quite buggy as it is. The purpose is to see if there is interest in such a feature and present the idea itself, not to merge it to master in its initial state.